requirements.txt
speedtest-cli==2.1.3
viam-sdk==0.42.0

typing-extensions
